all todo task completed, need to work on bulk order java now.

+++ Project2020/src/shop/ProductStock.java
@@ -1,6 +1,7 @@
 package shop;
 
 import java.util.Dictionary;
+import java.util.HashMap;
 import java.util.Hashtable;
 /**
  * 
@@ -9,8 +10,10 @@ import java.util.Hashtable;
  */
 public interface ProductStock {
 	// Initializing a Dictionary 
-    Dictionary<String, Double> shopStock = new Hashtable<String, Double>();
+    //HashMap<String, Integer> shopStock = new HashMap<String, Integer>();
+	Dictionary<String, Integer> shopStock = new Hashtable<String, Integer>();
     Dictionary<String, Double> shopPrice = new Hashtable<String, Double>();
+    //HashMap<String, Double> shopPrice = new HashMap<String, Double>();
